requirements.txt
# Runtime
pandas>=2.0
numpy>=1.24
pybit>=5.0
requests>=2.28
xgboost>=1.7

# Training
scikit-learn>=1.3

# Database — PostgreSQL (optional; SQLite is built-in)
psycopg2-binary>=2.9

# Read-only API + dashboard
fastapi>=0.110
uvicorn[standard]>=0.29
